Refactor #278 gmail 패키지 명 변경 및 코드 컨벤션 준수
