+++ src/main/java/com/gnims/project/domain/user/dto/PasswordDto.java
@@ -2,7 +2,6 @@ package com.gnims.project.domain.user.dto;
 
 import com.gnims.project.share.validation.ValidationGroups;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.Pattern;
@@ -10,9 +9,7 @@ import javax.validation.constraints.Pattern;
 import static com.gnims.project.share.message.ExceptionMessage.*;
 
 @Getter
-@NoArgsConstructor
 public class PasswordDto {
-
     @NotBlank(message = OLD_SECRET_EMPTY_MESSAGE,
             groups = ValidationGroups.NotNullGroup.class)
     @Pattern(regexp = "^(?=.*[a-zA-Z])(?=.*\\d)[a-zA-Z\\d]{8,16}$",
