+++ src/main/java/com/gnims/project/domain/user/dto/SimpleMessageResult.java
@@ -4,7 +4,7 @@ import lombok.Getter;
 
 @Getter
 public class SimpleMessageResult {
-    private Integer status;
+    private final Integer status;
     private final String message;
 
     public SimpleMessageResult(Integer status, String message) {
