+++ src/main/java/com/gnims/project/domain/user/repository/UserRepository.java
@@ -11,6 +11,7 @@ import java.util.List;
 import java.util.Optional;
 
 public interface UserRepository extends JpaRepository<User, Long> {
+
     Optional<User> findByEmail(String email);
     Optional<User> findByUsername(String username);
     Optional<User> findByNickname(String nickname);
