+++ src/main/java/com/gnims/project/scheduler/EmailScheduler.java
@@ -1,11 +1,12 @@
 package com.gnims.project.scheduler;
 
+import com.gnims.project.share.email.EmailRepository;
 import com.gnims.project.share.slack.SlackMessageSender;
-import com.gnims.project.share.gmail.EmailRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
+
 import java.time.LocalDate;
 import java.time.LocalDateTime;
