+++ src/main/java/com/gnims/project/share/email/EmailController.java
@@ -1,4 +1,4 @@
-package com.gnims.project.share.gmail;
+package com.gnims.project.share.email;
 
 import com.gnims.project.domain.user.dto.SimpleMessageResult;
 import com.gnims.project.share.validation.ValidationSequence;
@@ -17,22 +17,19 @@ import static org.springframework.http.ResponseEntity.ok;
 @RestController
 @RequiredArgsConstructor
 public class EmailController {
+
     private final EmailService emailService;
 
     @PatchMapping("/auth/password")
     public ResponseEntity<SimpleMessageResult> updatePassword(
             @Validated(ValidationSequence.class) @RequestBody EmailPasswordDto request)  throws Exception {
-
         emailService.updatePassword(request);
-
         return ok(new SimpleMessageResult(OK.value(), SECRET_UPDATE_SUCCESS_MESSAGE));
     }
 
     @PatchMapping("/auth/code")
     public ResponseEntity<SimpleMessageResult> updatePassword(@RequestBody AuthCodeDto request)  throws Exception {
-
         emailService.checkCode(request);
-
         return ok(new SimpleMessageResult(OK.value(), SUCCESS_AUTH_EMAIL_MESSAGE));
     }
 }
