+++ src/main/java/com/gnims/project/share/validation/ValidationGroups.java
@@ -1,6 +1,6 @@
 package com.gnims.project.share.validation;
 
 public class ValidationGroups {
-    public interface NotNullGroup {};
-    public interface PatternCheckGroup {};
+    public interface NotNullGroup {}
+    public interface PatternCheckGroup {}
 }
