+++ src/main/java/com/gnims/project/social/controller/SocialLoginController.java
@@ -17,13 +17,13 @@ import javax.servlet.http.HttpServletResponse;
 @RestController
 @RequiredArgsConstructor
 public class SocialLoginController {
+
     private final KakaoService kakaoService;
     private final NaverService naverService;
 
     @PostMapping ("/social/kakao-login")
     public ResponseEntity<SocialResult> kakaoLogin(@RequestBody SocialCodeDto codeDto,
                                                    HttpServletResponse response) throws JsonProcessingException {
-
         // code: 카카오 서버로부터 받은 인가 코드
         SocialResult result = kakaoService.kakaoLogin(codeDto.getCode(), response);
         return ResponseEntity.status(result.getStatus()).body(result);
@@ -32,7 +32,6 @@ public class SocialLoginController {
     @PostMapping ("/social/naver-login")
     public ResponseEntity<SocialResult> naverLogin(@RequestBody SocialTokenDto tokenDto,
                                                    HttpServletResponse response) throws JsonProcessingException {
-
         // token: 네이버 서버로부터 받은 사용자 정보 접근 토큰
         SocialResult result = naverService.naverLogin(tokenDto.getToken(), response);
         return ResponseEntity.status(result.getStatus()).body(result);
