+++ src/main/java/com/gnims/project/social/dto/SocialEmailDto.java
@@ -4,7 +4,7 @@ import lombok.Getter;
 
 @Getter
 public class SocialEmailDto {
-    private String email;
+    private final String email;
 
     public SocialEmailDto(String email) {
         this.email = email;
