+++ src/main/java/com/gnims/project/social/dto/SocialProfileDto.java
@@ -1,12 +1,10 @@
 package com.gnims.project.social.dto;
 
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 
 @Getter
-@NoArgsConstructor
 public class SocialProfileDto {
-    private String email;
+    private final String email;
 
     public SocialProfileDto(String email) {
         this.email = email;
