+++ src/test/java/com/gnims/project/domain/user/service/EmailCheckTest.java
@@ -36,7 +36,6 @@ public class EmailCheckTest {
 
     @BeforeEach
     void beforeEach() throws Exception {
-
         MockMultipartFile signupFile1 = new MockMultipartFile("data", "", "application/json", "{\"nickname\" : \"딸기\",\"username\": \"이땡땡\", \"email\": \"ddalgi@gmail.com\", \"password\": \"123456aA9\"}".getBytes());
 
         mvc.perform(multipart("/auth/signup").file(signupFile1).characterEncoding("utf-8"));
@@ -44,14 +43,12 @@ public class EmailCheckTest {
 
     @AfterEach
     void afterEach() throws Exception {
-
         userRepository.deleteAll();
     }
 
     @DisplayName("이메일 중복체크 성공 - 상태코드 200, 성공 메세지 반환")
     @Test
     void 중복체크성공테스트() throws Exception {
-
         mvc.perform(post("/auth/email")
                         .contentType(MediaType.APPLICATION_JSON)
                         .content("{\"email\": \"orange@gmail.com\"}"))
@@ -62,7 +59,6 @@ public class EmailCheckTest {
     @DisplayName("이메일 중복체크 null or 정규식 실패 - 상태코드 400, 실패 메세지 반환")
     @Test
     void 중복체크실패테스트1() throws Exception {
-
         mvc.perform(post("/auth/email")
                         .contentType(MediaType.APPLICATION_JSON)
                         .content("{}"))
@@ -79,7 +75,6 @@ public class EmailCheckTest {
     @DisplayName("이메일 중복체크 중복 실패 - 상태코드 400, 실패 메세지 반환")
     @Test
     void 중복체크실패테스트2() throws Exception {
-
         mvc.perform(post("/auth/email")
                         .contentType(MediaType.APPLICATION_JSON)
                         .content("{\"email\": \"ddalgi@gmail.com\"}"))
