+++ src/test/java/com/gnims/project/domain/user/service/UserSignupTest.java
@@ -48,7 +48,6 @@ public class UserSignupTest {
     @DisplayName("회원가입 성공(이미지 있을 때) - 상태코드 201, 성공 메세지를 반환, db에 저장")
     @Test
     void 회원가입성공테스트1() throws Exception {
-
         MockMultipartFile signupFile = new MockMultipartFile("data", "", "application/json", "{\"nickname\" : \"딸기\",\"username\": \"이땡땡\", \"email\": \"ddalgi@gmail.com\", \"password\": \"123456aA9\"}".getBytes());
 
         final String fileName = "르탄이"; //파일명
@@ -89,7 +88,6 @@ public class UserSignupTest {
     @DisplayName("회원가입 성공(이미지 null일 때) - 상태코드 201, 성공 메세지를 반환, db에 저장, 이미지는 기본 이미지 저장")
     @Test
     void 회원가입성공테스트2() throws Exception {
-
         MockMultipartFile signupFile = new MockMultipartFile("data", "", "application/json", "{\"nickname\" : \"딸기\",\"username\": \"이땡땡\", \"email\": \"ddalgi@gmail.com\", \"password\": \"123456aA9\"}".getBytes());
 
         //이미지 없을 때
@@ -108,7 +106,6 @@ public class UserSignupTest {
     @DisplayName("회원가입 성공, 토큰이 같이 왓을 시 - 상태코드 201, 성공 메세지를 반환, db에 저장")
     @Test
     void 회원가입토큰성공테스트() throws Exception {
-
         MockMultipartFile kakaoFile = new MockMultipartFile("data", "", "application/json", "{\"nickname\" : \"포도\", \"username\": \"김땡땡\", \"email\": \"orange@kakao.com\", \"password\": \"123456aA9\"}".getBytes());
 
         mvc.perform(multipart("/auth/signup")
@@ -123,7 +120,6 @@ public class UserSignupTest {
     @DisplayName("회원가입 시 닉네임 중복 - 상태코드 400, 에러 메세지를 반환, db에 저장 실패")
     @Test
     void 회원가입실패테스트1() throws Exception {
-
         //중복 체크를 위해 db에 넣을 데이터
         MockMultipartFile successFile = new MockMultipartFile("data", "", "application/json", "{\"nickname\" : \"딸기\",\"username\": \"이땡땡\", \"email\": \"ddalgi@gmail.com\", \"password\": \"123456aA9\"}".getBytes());
 
@@ -146,7 +142,6 @@ public class UserSignupTest {
     @DisplayName("회원가입 시 이메일 중복 - 상태코드 400, 에러 메세지를 반환, db에 저장 실패")
     @Test
     void 회원가입실패테스트2() throws Exception {
-
         //중복 체크를 위해 db에 넣을 데이터
         MockMultipartFile successFile = new MockMultipartFile("data", "", "application/json", "{\"nickname\" : \"딸기\",\"username\": \"이땡땡\", \"email\": \"ddalgi@gmail.com\", \"password\": \"123456aA9\"}".getBytes());
 
@@ -169,7 +164,6 @@ public class UserSignupTest {
     @DisplayName("회원가입 시 닉네임 값이 null 혹은 정규식 불 일치 - 상태코드 400, 에러 메세지를 반환, db에 저장 실패")
     @Test
     void 회원가입실패테스트3() throws Exception {
-
         //닉네임 null 값
         MockMultipartFile failFile1 = new MockMultipartFile("data", "", "application/json", "{\"username\": \"신땡땡\", \"email\": \"orange@gmail.com\", \"password\": \"1234aA5678\"}".getBytes());
 
@@ -194,7 +188,6 @@ public class UserSignupTest {
     @DisplayName("회원가입 시 이메일 값이 null 혹은 정규식 불 일치 - 상태코드 400, 에러 메세지를 반환, db에 저장 실패")
     @Test
     void 회원가입실패테스트4() throws Exception {
-
         //닉네임 null 값
         MockMultipartFile failFile1 = new MockMultipartFile("data", "", "application/json", "{\"nickname\" : \"짱구\", \"username\": \"신땡땡\", \"password\": \"1234aA5678\"}".getBytes());
 
@@ -219,7 +212,6 @@ public class UserSignupTest {
     @DisplayName("회원가입 시 이름 값이 null 혹은 정규식 불 일치 - 상태코드 400, 에러 메세지를 반환, db에 저장 실패")
     @Test
     void 회원가입실패테스트5() throws Exception {
-
         //이름 null 값
         MockMultipartFile failFile1 = new MockMultipartFile("data", "", "application/json", "{\"nickname\" : \"짱구\", \"email\": \"orange@gmail.com\", \"password\": \"1234aA5678\"}".getBytes());
 
@@ -244,7 +236,6 @@ public class UserSignupTest {
     @DisplayName("회원가입 시 비밀번호 값이 null 혹은 정규식 불 일치 - 상태코드 400, 에러 메세지를 반환, db에 저장 실패")
     @Test
     void 회원가입실패테스트6() throws Exception {
-
         //비밀번호 null 값
         MockMultipartFile failFile1 = new MockMultipartFile("data", "", "application/json", "{\"nickname\" : \"짱구\", \"username\": \"신땡땡\", \"email\": \"orange@gmail.com\"}".getBytes());
 
@@ -269,7 +260,6 @@ public class UserSignupTest {
     @DisplayName("회원가입 시 (여러) 값이 null 혹은 정규식 불 일치 - 상태코드 400, 에러 메세지를 반환, db에 저장 실패")
     @Test
     void 회원가입실패테스트7() throws Exception {
-
         //이메일, 닉네임, 이름, 비밀번호 중 null 값
         MockMultipartFile failFile3 = new MockMultipartFile("data", "", "application/json", "{}".getBytes());
